+++ config/gasSchedule.go
@@ -9,7 +9,7 @@ import (
 
 const GasValueForTests = 1
 
-var AsyncCallbackGasLockForTests = 100_000
+var AsyncCallbackGasLockForTests = uint64(100_000)
 
 // GasScheduleMap (alias) is the map for gas schedule
 type GasScheduleMap = map[string]map[string]uint64
@@ -732,5 +732,5 @@ func FillGasMap_WASMOpcodeValues(value uint64) map[string]uint64 {
 }
 
 func MakeGasMapForTests() GasScheduleMap {
-	return MakeGasMap(GasValueForTests, uint64(AsyncCallbackGasLockForTests))
+	return MakeGasMap(GasValueForTests, AsyncCallbackGasLockForTests)
 }
